ledger-on-memory: Use pre-execution in tests, not batching.
