+++ ledger/ledger-on-memory/src/test/suite/scala/com/daml/ledger/on/memory/InMemoryLedgerReaderWriterIntegrationSpec.scala
@@ -3,5 +3,46 @@
 
 package com.daml.ledger.on.memory
 
+import java.util.concurrent.Executors
+
+import com.daml.ledger.participant.state.kvutils.ParticipantStateIntegrationSpecBase
+import com.daml.ledger.participant.state.kvutils.ParticipantStateIntegrationSpecBase.ParticipantState
+import com.daml.ledger.participant.state.kvutils.api.KeyValueParticipantState
+import com.daml.ledger.participant.state.v1.{LedgerId, ParticipantId}
+import com.daml.ledger.resources.ResourceOwner
+import com.daml.ledger.validator.StateKeySerializationStrategy
+import com.daml.lf.engine.Engine
+import com.daml.logging.LoggingContext
+import com.daml.metrics.Metrics
+
+import scala.concurrent.ExecutionContext
+
 class InMemoryLedgerReaderWriterIntegrationSpec
-    extends InMemoryLedgerReaderWriterIntegrationSpecBase(enableBatching = true)
+    extends ParticipantStateIntegrationSpecBase(s"In-memory ledger/participant") {
+
+  override val isPersistent: Boolean = false
+
+  override def participantStateFactory(
+      ledgerId: LedgerId,
+      participantId: ParticipantId,
+      testId: String,
+      metrics: Metrics,
+  )(implicit loggingContext: LoggingContext): ResourceOwner[ParticipantState] =
+    for {
+      dispatcher <- dispatcherOwner
+      committerExecutionContext <- ResourceOwner
+        .forExecutorService(() => Executors.newCachedThreadPool())
+        .map(ExecutionContext.fromExecutorService)
+      readerWriter <- new InMemoryLedgerReaderWriter.Owner(
+        ledgerId = ledgerId,
+        participantId = participantId,
+        keySerializationStrategy = StateKeySerializationStrategy.createDefault(),
+        metrics = metrics,
+        dispatcher = dispatcher,
+        state = InMemoryState.empty,
+        engine = Engine.DevEngine(),
+        committerExecutionContext = committerExecutionContext,
+      )
+    } yield new KeyValueParticipantState(readerWriter, readerWriter, metrics)
+
+}
