--- ledger/ledger-on-memory/src/test/suite/scala/com/daml/ledger/on/memory/NonBatchedInMemoryLedgerReaderWriterIntegrationSpec.scala
@@ -1,7 +1,0 @@
-// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.ledger.on.memory
-
-class NonBatchedInMemoryLedgerReaderWriterIntegrationSpec
-    extends InMemoryLedgerReaderWriterIntegrationSpecBase(enableBatching = false) {}
