+++ ledger/recovering-indexer-integration-tests/src/test/suite/scala/com/digitalasset/platform/indexer/RecoveringIndexerIntegrationSpec.scala
@@ -14,12 +14,10 @@ import akka.stream.scaladsl.Source
 import ch.qos.logback.classic.Level
 import com.codahale.metrics.MetricRegistry
 import com.daml.ledger.on.memory
-import com.daml.ledger.participant.state.kvutils.api.{
-  BatchingLedgerWriterConfig,
-  KeyValueParticipantState,
-}
+import com.daml.ledger.participant.state.kvutils.api.KeyValueParticipantState
 import com.daml.ledger.participant.state.v1._
 import com.daml.ledger.resources.{ResourceOwner, TestResourceContext}
+import com.daml.ledger.validator.StateKeySerializationStrategy
 import com.daml.lf.data.Ref
 import com.daml.lf.data.Ref.LedgerString
 import com.daml.lf.engine.Engine
@@ -256,13 +254,22 @@ object RecoveringIndexerIntegrationSpec {
         loggingContext: LoggingContext,
     ): ResourceOwner[ParticipantState] = {
       val metrics = new Metrics(new MetricRegistry)
-      new memory.InMemoryLedgerReaderWriter.SingleParticipantBatchingOwner(
-        ledgerId,
-        BatchingLedgerWriterConfig.reasonableDefault,
-        participantId,
-        metrics = metrics,
-        engine = Engine.DevEngine(),
-      ).map(readerWriter => new KeyValueParticipantState(readerWriter, readerWriter, metrics))
+      for {
+        dispatcher <- memory.dispatcherOwner
+        committerExecutionContext <- ResourceOwner
+          .forExecutorService(() => Executors.newCachedThreadPool())
+          .map(ExecutionContext.fromExecutorService)
+        readerWriter <- new memory.InMemoryLedgerReaderWriter.Owner(
+          ledgerId = ledgerId,
+          participantId = participantId,
+          keySerializationStrategy = StateKeySerializationStrategy.createDefault(),
+          metrics = metrics,
+          dispatcher = dispatcher,
+          state = memory.InMemoryState.empty,
+          engine = Engine.DevEngine(),
+          committerExecutionContext = committerExecutionContext,
+        )
+      } yield new KeyValueParticipantState(readerWriter, readerWriter, metrics)
     }
   }
